Quyen

+++ mavenproject2/src/main/java/com/pojos/Cart.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 public class Cart {
     private int productId;
     private String name;
-    private BigDecimal price;
+    private Long price;
     private int count;
     private String image;
 
@@ -49,14 +49,14 @@ public class Cart {
     /**
      * @return the price
      */
-    public BigDecimal getPrice() {
+    public Long getPrice() {
         return price;
     }
 
     /**
      * @param price the price to set
      */
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
